profile argument added to commands
