+++ packages/cli/src/command-implementations/deploy-command.ts
@@ -22,13 +22,6 @@ export class DeployCommand extends ProjectCommand<DeployCommandDefinition> {
 
     const projectDef = await this.readProjectDefinitionFromFS()
 
-    if (this.argv.profile) {
-      const config = await this.readProfileFromFS(this.argv.profile)
-      this.argv.apiUrl = config.apiUrl
-      this.argv.workspaceId = config.workspaceId
-      this.argv.token = config.token
-    }
-
     if (projectDef.type === 'integration') {
       return this._deployIntegration(api, projectDef.definition)
     }
